rename to ISocket
